Changed `bs-platform` from "dependencies" to "peerDependencies"
It is a good practice to add 'bs-platform' as "peerDependency" so it doesn't trigger rebuilding bucklescript for every package, which would be slow and undesirable. The standard practice in Reason is to 'link' the global bs-platform package once per project instead so only a single `bs-platform` package is used without rebuilding it. People using Reason are used to this.